requirements.txt
wagtail
python-dotenv
django-crispy-forms
crispy-bootstrap5